+++ philo_bonus/include/utils.h
@@ -3,7 +3,7 @@
 
 bool		atoi_strict(char *str, int *dst);
 long long	get_usec(void);
-int			max(int a, int b);
+long long	max(long long a, long long b);
 void		usleep_until(long long time);
 void		ft_putstr_fd(const char *s, int fd);
